refactor: `cargo-miden` to handle only it's own commands (new, build, example)
and move build-related code to the `BuildCommand`.
