+++ tools/cargo-miden/src/commands/mod.rs
@@ -1,5 +1,7 @@
+pub mod build;
 pub mod example_project;
 pub mod new_project;
 
+pub use build::BuildCommand;
 pub use example_project::ExampleCommand;
 pub use new_project::NewCommand;
